% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metareport.R
\name{metareport_summary}
\alias{metareport_summary}
\title{summary version of metareport}
\usage{
metareport_summary(
  data_file,
  meta_file = NULL,
  output_format = "html_document"
)
}
\arguments{
\item{data_file}{}

\item{meta_file}{}

\item{output_format}{}
}
\value{

}
\description{
summary version of metareport
}
