% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_peptides_file.R
\name{render_peptides_file}
\alias{render_peptides_file}
\title{wrapper function of rmarkdown::render for opencpu to render a report for peptides.txt from markdown template}
\usage{
render_peptides_file(
  file,
  meta = NULL,
  template_version = "stable",
  output_format = "html_document",
  output_dir = NULL
)
}
\arguments{
\item{file}{string, the file (path) of the peptide.txt, from maxquant result txt folder mandatory}

\item{meta}{the file (path), optional, string is the tsv file, generated by MetaLab2.0, 1st columns as sample name, 2nd column as experiment name, 3rd column and after as grouping}

\item{template_version}{string, can be "latest","stable", or a specific number, like "1.0", "1.1". Default as "stable". Will use the local template in the package (system.file("rmd","MQ_report_summary.Rmd", package = "metareport")), otherwise, it will try to curl the corresponding file from the deposit from either gitlab/github. Besure to use the correct number. If the version defined does not exist, or failed to retrieve from the deposit, it will use the stable version instead.}

\item{output_format}{see output_format in render}

\item{output_dir}{see output_dir in render}
}
\value{
no direct return, but write an output.html to the temp session on the opencpu server
}
\description{
This function is destined for working on the opencpu server end, accepting file posted by front end to produce an html file to send to the front end.
It also works in  stand-alone mode, to generate a report file, output.html, using a public accessible rmd file on github. Of course, this need internet connection to run
}
\examples{
  datafile <- system.file("extdata","peptide.txt", package = "metareport")
  metafile <-system.file("extdata","metainfo.txt", package = "metareport")
  
  # generate a report without meta information 
  render_peptide_file(file = datafile, output_dir =  getwd())
  # generate a report with meta information
}
\seealso{
\code{\link{render}}  \code{\link{knit}} \code{\link{render_MQsummary}} \code{\link{render_MQsummary_file}} \code{\link{render_proteinGroups_file}}
}
