% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd_server.R
\name{render_rmd_server}
\alias{render_rmd_server}
\title{A wrapper function of rmarkdown::render for opencpu}
\usage{
render_rmd_server(rmd_file)
}
\arguments{
\item{rmd_file}{a string of the file name on the same opencpu server in the same metareport package,}
}
\value{
no direct return, but write an output.html to the temp session on the opencpu server
}
\description{
A basic wrapper function to work with opencpu.
}
\details{
render_rmd_server will render a local rmd file into html 

core function,  only for debuging and testing
}
\examples{
  path_to_input <- system.file("rmd", rmd_file, package = "rmdocpu")
  render_rmd_server(path_to_input)

}
\seealso{
\code{\link{render}}  \code{\link{knit}}
}
