% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_taxon_file.R
\name{render_taxon_file}
\alias{render_taxon_file}
\title{wrapper function of rmarkdown::render for opencpu to render a report for MetaLab_taxonomy.xlsx}
\usage{
render_taxon_file(file, meta = NULL)
}
\arguments{
\item{file}{the file path, string is the tsv file, taxon.txt}

\item{meta}{the file (path), optional, string is the tsv file, generated by MetaLab2.0, 1st columns as sample name, 2nd column as experiment name, 3rd column and after as grouping}
}
\value{
no direct return, but write an output.html to the temp session on the opencpu server
}
\description{
This function works on the opencpu server end, to produce an html file, which can then be retrieved by javascript or any other customer api tool.
It also works on the stand alone mode, to generate a report file, output.html in the currrent folder, using a public accessible rmd file on github.
Of course, this need internect connection to run
}
\examples{
  render_taxon_file(file = "MetaLab_taxonomy.xlsx")
  render_taxon_file(file = "MetaLab_taxonomy.xlsx", meta = "meta.txt")
}
\seealso{
\code{\link{render}}  \code{\link{knit}} \code{\link{render_MQsummary}} \code{\link{render_MQsummary_file}} \code{\link{render_proteinGroups_file}}
}
